Current changes

+++ zebrazoom/mainZZ.py
@@ -19,6 +19,7 @@ import cv2
 import zebrazoom.videoFormatConversion.zzVideoReading as zzVideoReading
 import json
 import subprocess
+import glob
 
 try:
   from PyQt6.QtWidgets import QApplication
@@ -125,7 +126,13 @@ def mainZZ(pathToVideo, videoName, videoExt, configFile, argv):
     # Creating output folder
     if not(hyperparameters["reloadWellPositions"]) and not(hyperparameters["reloadBackground"]) and not(os.path.exists(os.path.join(outputFolderVideo, 'intermediaryWellPositionReloadNoMatterWhat.txt'))) and not(hyperparameters["dontDeleteOutputFolderIfAlreadyExist"]):
       if os.path.exists(outputFolderVideo):
-        shutil.rmtree(outputFolderVideo)
+        if glob.glob(os.path.join(outputFolderVideo, 'results_*.txt')) and glob.glob(os.path.join(outputFolderVideo, '*.avi')):
+          pastNumbersTaken = 1
+          while os.path.exists(outputFolderVideo + '_PastTracking_' + str(pastNumbersTaken)) and pastNumbersTaken < 99:
+            pastNumbersTaken += 1
+          shutil.move(outputFolderVideo, outputFolderVideo + '_PastTracking_' + str(pastNumbersTaken))
+        else:
+          shutil.rmtree(outputFolderVideo)
       if hyperparameters["tryCreatingFolderUntilSuccess"]:
         while True:
           try:
